Move Module list in a dedicated file (#445)
