+++ app/api.py
@@ -1,21 +1,18 @@
 """File defining all the routes for the module, to configure the router"""
 
-import importlib
 import logging
-from pathlib import Path
 
 from fastapi import APIRouter
 
 from app.core import endpoints_core
 from app.core.auth import endpoints_auth
 from app.core.groups import endpoints_groups
-from app.core.module import Module
 from app.core.notification import endpoints_notification
 from app.core.users import endpoints_users
+from app.modules.module_list import module_list
 
 hyperion_error_logger = logging.getLogger("hyperion.error")
 
-module_list: list[Module] = []
 api_router = APIRouter()
 
 api_router.include_router(endpoints_auth.router)
@@ -24,15 +21,5 @@ api_router.include_router(endpoints_groups.router)
 api_router.include_router(endpoints_notification.router)
 api_router.include_router(endpoints_users.router)
 
-for endpoints_file in Path().glob("app/modules/*/endpoints_*.py"):
-    endpoint_module = importlib.import_module(
-        ".".join(endpoints_file.with_suffix("").parts),
-    )
-    if hasattr(endpoint_module, "module"):
-        module: Module = endpoint_module.module
-        module_list.append(module)
-        api_router.include_router(module.router)
-    else:
-        hyperion_error_logger.error(
-            f"Module {endpoints_file} does not declare a module. It won't be enabled.",
-        )
+for module in module_list:
+    api_router.include_router(module.router)
